+++ src/errors/js.ts
@@ -33,6 +33,7 @@ class JSErrors extends Base {
         col,
         message,
         collector: options.collector,
+        stack: error.stack,
       };
       this.traceInfo();
     };