+++ CHANGES.md
@@ -1,8 +1,16 @@
 # Changes
 
+## 0.4.0
+
+1. Update stack and message in logs.
+2. Fix wrong URL when using relative path in xhr.
+
 ## 0.3.0
 
 1. Support tracing starting at the browser.
+2. Add traceSDKInternal SDK for tracing SDK internal RPC.
+3. Add detailMode SDK for tracing http method and url as tags in spans.
+4. Fix conditions of http status.
 
 ## 0.2.0